requirements.txt
flask
mysql-connector-python
python-dotenv
werkzeug
pandas
h2o
customtkinter
tkcalendar
matplotlib
numpy
